backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
decorator==4.4.2
deep-translator==1.11.4
distro==1.9.0
elevenlabs==2.7.1
fastapi==0.116.0
filelock==3.18.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fpdf==1.7.2
frozenlist==1.6.2
fsspec==2025.5.1
google-ai-generativelanguage==0.4.0
google-api-core==2.25.0
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.27.0
google-genai==1.22.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
googletrans==4.0.2
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.62.3
gTTS==2.5.4
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
language_tool_python==2.9.4
livereload==2.7.1
lxml==5.4.0
MarkupSafe==3.0.2
MetaTrader5==5.0.5120
midtransclient==1.4.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.4.4
mutagen==1.47.0
mysql==0.0.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
networkx==3.5
numpy==2.2.6
openai==1.86.0
opencv-python==4.11.0.86
packaging==25.0
pefile==2023.2.7
Pillow==10.0.1
proglog==0.1.12
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
python-docx==1.1.2
python-dotenv==1.0.0
python-multipart==0.0.20
pywin32-ctypes==0.2.3
replicate==1.0.7
reportlab==4.4.1
requests==2.31.0
pytesseract==0.3.10
scikit-learn==1.3.0
rfc3986==1.5.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
yt-dlp==2025.6.9
waitress==2.1.2